+++ shianghergo/src/main/java/com/shianghergo/dao/OrderDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.shianghergo.model.OrderBean;
 
 public interface OrderDao {
-	List<OrderBean> getOrderBeanByMemeber(Long member_id);
-	Long addOrder(Long member_id);
-	void updateStatus(Long order_id);
+	List<OrderBean> getOrderBeanByMemeber(Integer member_id);
+	Integer addOrder(Integer member_id);
+	void updateStatus(Integer order_id);
 }
