+++ shianghergo/src/main/java/com/shianghergo/dao/OrderDetailDao.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.shianghergo.model.OrderDetailBean;
 
 public interface OrderDetailDao {
-	List<OrderDetailBean> getOrderDetail(Long id);
+	List<OrderDetailBean> getOrderDetail(Integer id);
 }
