+++ shianghergo/src/main/java/com/shianghergo/model/OrderDetailBean.java
@@ -9,38 +9,38 @@ import javax.persistence.Table;
 @Entity
 @Table(name="order_detail")
 public class OrderDetailBean {
-	private Long id;
-	private Long order_id;
-	private Long store_id;
-	private Long item_id;
+	private Integer id;
+	private Integer order_id;
+	private Integer store_id;
+	private Integer item_id;
 	private String name;
-	private Long price;
+	private Integer price;
 	private Integer amount;
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
-	public Long getOrder_id() {
+	public Integer getOrder_id() {
 		return order_id;
 	}
-	public void setOrder_id(Long order_id) {
+	public void setOrder_id(Integer order_id) {
 		this.order_id = order_id;
 	}
-	public Long getStore_id() {
+	public Integer getStore_id() {
 		return store_id;
 	}
-	public void setStore_id(Long store_id) {
+	public void setStore_id(Integer store_id) {
 		this.store_id = store_id;
 	}
-	public Long getItem_id() {
+	public Integer getItem_id() {
 		return item_id;
 	}
-	public void setItem_id(Long item_id) {
+	public void setItem_id(Integer item_id) {
 		this.item_id = item_id;
 	}
 	public String getName() {
@@ -49,10 +49,10 @@ public class OrderDetailBean {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Long getPrice() {
+	public Integer getPrice() {
 		return price;
 	}
-	public void setPrice(Long price) {
+	public void setPrice(Integer price) {
 		this.price = price;
 	}
 	public Integer getAmount() {
