+++ shianghergo/src/main/java/com/shianghergo/service/OrderService.java
@@ -6,7 +6,7 @@ import com.shianghergo.model.OrderBean;
 
 public interface OrderService {
 	
-	List<OrderBean> getOrderBeanByMemeber(Long member_id);
-	Long addOrder(Long member_id);
-	void updateStatus(Long order_id);
+	List<OrderBean> getOrderBeanByMemeber(Integer member_id);
+	Integer addOrder(Integer member_id);
+	void updateStatus(Integer order_id);
 }
